requirements.txt
beautifulsoup4==4.5.1
bokeh==0.12.2
chardet==2.3.0
cssselect==0.9.2
dateutil
elasticsearch-dsl==2.1.0
elasticsearch==2.3.0
eventlet
flask
Flask-CORS
flask_jsonrpc
gensim
gevent
HTMLParser
httplib2
javaobj
lxml
matplotlib
networkx
numpy
objgraph
pandas
paramiko
Pillow
portalocker
psutil
pymongo
pymysql
pytools
pywinauto
redis
requests
slackclient
sqlalchemy
tabulate
xlrd
xlwt
